Update question, response and expected results (question 3)
As a result of recently requested changes that requires addition of two new
options for new and used cars.

I have updated the simplified-expenses-checker-questions-and-responses.yml and
simplified-expenses-checker-responses-and-expected-results.yml files with the
aforementioned options.

This will allow for the generation of regression artefacts before and after the
proposed  changes, showing the difference after the changes are made.

This commit also shows the affected/changes to regression test artefacts as a
result of the addition of options for new and used cars to the the
simplified-expenses-checker-questions-and-responses.yml and
simplified-expenses-checker-responses-and-expected-results.yml files.

I have reviewed the changes, though lengthy but it is as expected.

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned changes have had no (interfering) affect to
other smart answers.